Add files via upload
The objective of this practical is to develop a Python program that 
generates a random password of a specified length. This program uses the random 
module to generate characters and the string module to get a possible characters 
(uppercase letters, lowercase letters, digits and special characters). This exercise will 
help students understand basic programming concepts like creates a password by 
randomly selecting characters from this set. The main function takes user input for the 
password length and then calls to display the generated password.